Test using programmatic compilation
Rather than cover code using the it-project (which is difficult to
organize and cumbersome to write tests with), use the Java compiler API
to programmatically compile sources generated by Lily in lily-compiler
test cases. The major advantage of this approach is that we can define
small OAS documents as strings within java test cases, keeping every
scenario isolated from the others.