Refactor schema evaluation for re-use during param evaluation
The original draft assumed that we needed to generate AstClassAlias instances, which is exclusively necessary during Component evaluation only. The new draft does not create aliases, but returns enough information about evaluated AST that another layer of the compiler *can* create those aliases and modify AstReferences as necessary. That responsibility now belongs to the OasComponentsToAst module.

This will let us re-use schema generation on request parameters, which is the motivation of this refactoring.

This is a WIP. Everything works, but the code is rough-edged (commented out a few test cases, though I think they're redundant anyways.)