broken|wip - almost finish sealed interfaces
This has one-half of the circular relationship between the iface and permitted members. In order to go both ways, we need to then modify the generated classes with an extends clause. I think the OasSchemaToAst should return the generated AstClass instances rather than refs (since a ref is just a fqn anyways!), which would let us customize generated code further. So, refactoring next?