Issues/27 Rename all/other tag groups (#61)
Previously, every operation was automatically a member of the 'all'
tag, and operations without tags became members of the 'other' tag. This
created a potential collision if the API actually used either of those
tags itself, since 'allOperations' could denote either operations with
the 'all' tag from the document or the special-case group containing
every known operation.

We now treat the all and other groups as special cases. These manifest
as the everyOperation and everyUntaggedOperation groups, whose names
cannot conflict with ${tag}Operations groups due to the differing naming
conventions.