Add warning documentation to empty classes
When Lily is unable to generate a class due to unsupported keywords, it
adds a Javadoc to the class (in addition to log warnings) reminding the
user that it was unable to create a useful class.

The class is generated so that references to the FQN still resolve at
compile time, which avoids the headache of post-processing the graph of
references.

Co-authored-by: hdcpereira <hdcpereira@protonmail.com>
Co-authored-by: Luke Green <lukeg496@gmail.com>