Add OneOf support for $ref component
Creates a new interface for the OneOf schema, and adds the implements
clause to the $ref class using a new Modifier, AddInterface.

This changes the core from a pure uninterrupted stream to a two-step
process: The first step collects Definitions and Modifiers together into
a table of name -> definition, then streams the finished result out for
rendering. As a result, the whole AST is collected in memory.

Co-authored-by: (Jonathan) Luke Green <lukeg496@gmail.com>
Co-authored-by: Hugo Pereira <hdcpereira@protonmail.com>