Add default revision to all poms
If we run exec:java with a -pl list of modules, it will fail to resolve
dependencies on othe rlily modules because exec:java doesn't resolve
$revision using the value form the parent pom. We've added a default
value to all modules so they can always resolve the version.

The version is also set to 0.0.0-SNAPSHOT instead of just SNAPSHOT so
that maven stops logging a massive warning.