Use jackson to deserialize OAS document
We make aggressive use of Optionals and None-types to prevent any null
pointers from sneaking in. This has exposed a large number of places
where we forgot to check for nulls, which have been annotated with TODOs
for now.

Jackson has been customized to skip over unparseable fields whenever
possible, so if the specification contains e.g. an erroneous $ref field
in a parameter definition and nothing else, jackson will just skip that
element rather than unroll the stack.