Remove .path from operation if there are no path parameters
Discovered this while creating a second clojure test case, so I fixed
it.

Also refactoring the clojure tests a bit. I've removed imports for
generated code -- they weren't necessary and caused issues. We're also
now able to specify the OpenAPI spec as a hash-map, which jackson
converts to an OpenAPI object like a champ. This lets us write helpers
for quickly setting up permutations of the same test, which is a
delight.